round_p.c: fixed types for fprintf when WANT_ASSERT >= 2.
[From SVN r5774 (trunk)]

+++ round_p.c
@@ -37,8 +37,8 @@ mpfr_round_p (mp_limb_t *bp, mp_size_t bn, mp_exp_t err0, mp_prec_t prec)
   if (i1 != i2)
     {
       fprintf (stderr, "mpfr_round_p(%d) != mpfr_can_round(%d)!\n"
-              "bn = %ld, err0 = %ld, prec = %ld\nbp = ",
-              i1, i2, bn, err0, prec);
+               "bn = %lu, err0 = %ld, prec = %ld\nbp = ",
+               i1, i2, (unsigned long) bn, (long) err0, (long) prec);
       gmp_fprintf (stderr, "%NX\n", bp, bn);
       MPFR_ASSERTN (0);
     }
